requirements.txt
attrs>=19.3.0
eventlet>=0.25.2
Flask>=1.1.2
Flask-Cors>=3.0.8
Flask-SocketIO>=4.3.0
gevent>=20.6.2
gevent-websocket>=0.10.1
gunicorn>=20.0.4
Jinja2>=2.11.2
python-engineio>=3.13.0
python-socketio>=4.6.0
Werkzeug>=1.0.1